Move QuerySanitizer under the Search namespace
